Gave files better names
note that currently the heat-stack-template.yaml actually includes the
cloud init yaml files ci-master.yaml and ci-slave.yaml. Actually an
updated version of these files. The next step will be to allow inclusion
of these file rather than directly containing the content to allow for
easier parsing and reading of the files.